resource/aws_autoscaling_group: Implement keyvaluetags and support provider ignore_tags
Reference: https://github.com/terraform-providers/terraform-provider-aws/issues/13808

Output from acceptance testing:

```
--- PASS: TestAccAWSAutoScalingGroup_ALB_TargetGroups (179.86s)
--- PASS: TestAccAWSAutoScalingGroup_ALB_TargetGroups_ELBCapacity (333.20s)
--- PASS: TestAccAWSAutoScalingGroup_autoGeneratedName (51.77s)
--- PASS: TestAccAWSAutoScalingGroup_basic (240.41s)
--- PASS: TestAccAWSAutoScalingGroup_classicVpcZoneIdentifier (134.91s)
--- PASS: TestAccAWSAutoScalingGroup_enablingMetrics (177.09s)
--- PASS: TestAccAWSAutoScalingGroup_initialLifecycleHook (243.57s)
--- PASS: TestAccAWSAutoScalingGroup_launchTemplate (45.88s)
--- PASS: TestAccAWSAutoScalingGroup_LaunchTemplate_IAMInstanceProfile (58.72s)
--- PASS: TestAccAWSAutoScalingGroup_launchTemplate_update (106.61s)
--- PASS: TestAccAWSAutoScalingGroup_launchTempPartitionNum (51.25s)
--- PASS: TestAccAWSAutoScalingGroup_LoadBalancers (618.13s)
--- PASS: TestAccAWSAutoScalingGroup_MaxInstanceLifetime (165.12s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy (44.59s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_InstancesDistribution_OnDemandAllocationStrategy (42.10s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_InstancesDistribution_OnDemandBaseCapacity (76.16s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_InstancesDistribution_OnDemandPercentageAboveBaseCapacity (47.32s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_InstancesDistribution_SpotAllocationStrategy (43.62s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_InstancesDistribution_SpotInstancePools (82.44s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_InstancesDistribution_SpotMaxPrice (150.94s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_InstancesDistribution_UpdateToZeroOnDemandBaseCapacity (78.70s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_LaunchTemplate_LaunchTemplateSpecification_LaunchTemplateName (45.47s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_LaunchTemplate_LaunchTemplateSpecification_Version (84.44s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_LaunchTemplate_Override_InstanceType (76.53s)
--- PASS: TestAccAWSAutoScalingGroup_MixedInstancesPolicy_LaunchTemplate_Override_WeightedCapacity (207.67s)
--- PASS: TestAccAWSAutoScalingGroup_namePrefix (50.25s)
--- PASS: TestAccAWSAutoScalingGroup_serviceLinkedRoleARN (54.73s)
--- PASS: TestAccAWSAutoScalingGroup_suspendingProcesses (198.11s)
--- PASS: TestAccAWSAutoScalingGroup_tags (232.54s)
--- PASS: TestAccAWSAutoScalingGroup_TargetGroupArns (247.63s)
--- PASS: TestAccAWSAutoScalingGroup_terminationPolicies (105.39s)
--- PASS: TestAccAWSAutoScalingGroup_VpcUpdates (175.27s)
--- PASS: TestAccAWSAutoScalingGroup_WithLoadBalancer (251.24s)
--- PASS: TestAccAWSAutoScalingGroup_WithLoadBalancer_ToTargetGroup (392.43s)
--- PASS: TestAccAWSAutoScalingGroup_withMetrics (81.00s)
--- PASS: TestAccAWSAutoScalingGroup_withPlacementGroup (146.95s)
```
